// features/rabbitmq/types.ts
export type QueueNodeData = {
  id: string;
  name: string;
  type: string;
  stats: {
    publish: number;
    ack: number;
  };
  consumer: {
    count: number;
    capacity: number;
  };
};